Anuj/Assignment 4/requirements.txt
fastapi==0.115.11
feedparser==6.0.11
Pillow==11.1.0
psycopg2_binary==2.9.10
python_dateutil==2.8.2
Requests==2.32.3
uvicorn==0.34.0
jinja2